Removing InfluenceField, getInfluence is on AbstractRobot
